Update and rename Install.sh to SLS101.sh
Alpha comin...SuperLazyScript 1.01...BETA, there are issues on some terminals, given the fact that this uses mate-terminal. There may be an issue on frontend, relating to zsh mate-terminal conflicts/technical changes that MAY cause this script to not work properly on your terminal. 
This was made in a VirtualBox, Kernel: 5.7.0-2parrot2-amd64 - Shell: zsh 5.8 - DE: MATE 1.24.1 - Terminal: mate-terminal.